requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
coveralls==1.7.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
dj-database-url==0.5.0
Django==2.2.1
django-heroku==0.3.1
djangorestframework==3.9.4
drf-yasg==1.12.1
django-extensions==2.1.7
entrypoints==0.3
factory-boy==2.12.0
flake8==3.7.7
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
isort==4.3.19
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
mccabe==0.6.1
pep8==1.7.1
psycopg2==2.7.7
psycopg2-binary==2.8.2
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pylint==2.3.1
pypi==2.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
urllib3==1.24.3
whitenoise==4.1.2
wrapt==1.11.1
